add propTypes, jsDoc, rename components
