+++ package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "axios": "^0.16.2",
     "immutable": "^3.8.1",
+    "prop-types": "^15.5.10",
     "react": "^15.5.4",
     "react-dom": "^15.5.4",
     "react-redux": "^5.0.5",
